\name{amoc-package}
\alias{amoc-package}
\alias{amoc}
\docType{package}
\title{
\packageTitle{amoc}
}
\description{
\packageDescription{amoc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{amoc}
\packageIndices{amoc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{amoc}

Maintainer: \packageMaintainer{amoc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
